updated Dockerfile
